# public/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Next REST Framework",
    "description": "This is an autogenerated documentation by Next REST Framework.",
    "version": "v5.1.11"
  },
  "paths": {
    "/api/author": {
      "post": {
        "operationId": "createAuthor",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAuthorRequestBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Response for status 201",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateAuthor201ResponseBody"
                }
              }
            }
          },
          "401": {
            "description": "Response for status 401",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateAuthor401ResponseBody"
                }
              }
            }
          },
          "422": {
            "description": "Response for status 422",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateAuthor422ResponseBody"
                }
              }
            }
          },
          "500": {
            "description": "Response for status 500",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateAuthor500ResponseBody"
                }
              }
            }
          }
        },
        "tags": ["author"]
      }
    }
  },
  "components": {
    "schemas": {
      "CreateAuthor201ResponseBody": {
        "type": "object",
        "properties": { "id": { "type": "number" } },
        "additionalProperties": false
      },
      "CreateAuthor401ResponseBody": { "type": "string" },
      "CreateAuthor422ResponseBody": { "type": "string" },
      "CreateAuthor500ResponseBody": {
        "type": "object",
        "properties": {},
        "additionalProperties": false
      },
      "CreateAuthorRequestBody": {
        "type": "object",
        "properties": {
          "id": { "type": "number" },
          "name": { "type": "string" },
          "language": { "type": "string" },
          "description": { "type": "string" }
        },
        "required": ["name", "language", "description"],
        "additionalProperties": false
      },
      "UnexpectedError": {
        "type": "object",
        "properties": { "message": { "type": "string" } },
        "additionalProperties": false
      }
    }
  }
}